requirements.txt
future~=0.18.2
Pillow~=8.0.1
praw~=7.1.0
requests~=2.24.0
opencv-python~=4.4.0.44
numpy~=1.19.2
Send2Trash~=1.5.0